Crate oku

Structs

CACHE_DIR

The platform-specific directory where Oku caches data

DATA_DIR

The platform-specific directory where Oku stores user data

DownloadItem

PICTURES_DIR

The platform-specific directory where users store pictures

PROJECT_DIRECTORIES

The platform-specific directories intended for Oku’s use

USER_DIRECTORIES

The platform-specific directories containing user files

Constants

VERSION

The current release version number of Oku

Functions

connect

Connect to a page using the current tab

create_initial_tab

Create an initial tab, for when the notebook is empty

download_ipfs_file_from_api

Download an IPFS file to the local machine using an existing IPFS node

download_ipfs_file_natively

Download an IPFS file to the local machine using an in-memory IPFS node

from_hash_natively

Get an IPFS file asynchronously using an in-memory IPFS node

from_hash_using_api

Get an IPFS file asynchronously using an existing IPFS node

handle_ipfs_request_natively

Comply with a request using the IPFS scheme natively

handle_ipfs_request_using_api

Comply with a request using the IPFS scheme

initial_connect

Perform the initial connection at startup when passed a URL as a launch argument

ipfs_options

main

The main function of Oku

new_about_dialog

Create a dialog box showing information about Oku

new_tab

Create a tab to be placed in the notebook

new_tab_label

Create the text to be displayed on a tab

new_tab_page

Create a new entry in the notebook

new_view

Create a new WebKit instance for the current tab

new_window

Create a new functional & graphical browser window

setup_native_ipfs

Setup an IPFS node

update_favicon

Update a tab’s icon

update_load_progress

Update the load progress indicator under the navigation bar

update_nav_bar

Update the contents of the navigation bar

update_title

Update a tab’s title

view

Get the WebKit instance for the current tab