Crate oku

Structs

CACHE_DIR
The platform-specific directory where Oku caches data

DATA_DIR
The platform-specific directory where Oku stores user data

DownloadItem

PICTURES_DIR
The platform-specific directory where users store pictures

PROJECT_DIRECTORIES
The platform-specific directories intended for Oku's use

USER_DIRECTORIES
The platform-specific directories containing user files

Constants

VERSION
The current release version number of Oku

Functions

connect
Connect to a page using the current tab

create_initial_tab
Create an initial tab, for when the notebook is empty

download_ipfs_file
Download an IPFS file to the local machine

get_from_hash
Get an IPFS file asynchronously

get_view
Get the WebKit instance for the current tab

handle_ipfs_request
Comply with a request using the IPFS scheme

main
The main function of Oku

new_about_dialog

new_tab
Create a tab to be placed in the notebook

new_tab_label
Create the text to be displayed on a tab

new_tab_page
Create a new entry in the notebook

new_view
Create a new WebKit instance for the current tab

new_window
Create a new functional & graphical browser window

update_favicon
Update a tab's icon

update_load_progress
Update the load progress indicator under the navigation bar

update_nav_bar
Update the contents of the navigation bar

update_title
Update a tab's title